backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytest
